==> 01-kza-cadastre-data-harvesting.Rmd <==
---
title: "Scrapping Kazakhstan cadastre data"
author: "Eduard Bukin"
date: "`r Sys.Date()`"
output:
  html_document: default
  pdf_document: default
editor_options:
  chunk_output_type: console
---



```{r setup, echo = FALSE, warning=FALSE}
library(pacman)
pacman::p_load(sf, here, fs, tidyverse, stringi, jsonlite, 
               httr, RCurl, progress, rgdal, glue, progressr)
force_reharvest <- FALSE
# knitr::opts_chunk(warning=FALSE, message = FALSE)
```



# Overview of the scrapping process

We aim to hares GIS polygons of land plots form the cadastre map of Kazakhstan:

-  [http://www.aisgzk.kz/aisgzk/ru/content/maps/](http://www.aisgzk.kz/aisgzk/ru/content/maps/)

This is a challenging task, because when we zoom in to the level of one rayon, we receive all maps in raters.

```{r out.width="100%", eval=TRUE, echo=FALSE}
smpl_poly_rast <- here("data-raw", "imgs", "sample-poly-rasters.png")

if (!file.exists(smpl_poly_rast)) {
  "http://www.aisgzk.kz/aisgzk/Proxy/aisgzkZem2/MapServer/export?dpi=96&transparent=true&format=png8&layers=show%3A147&bbox=7942828.607889999%2C6208598.464414631%2C8243720.20321%2C6512657.339685368&bboxSR=3857&imageSR=3857&size=760%2C768&_ts=1616703947839&layerDefs=147%3AShape_Area%20%3E%3D%20748174&f=image" %>% 
  download.file(smpl_poly_rast)
}

knitr::include_graphics(smpl_poly_rast)
```

We also receive points of the coordinates centroids in raster format.

```{r out.width="100%", eval=TRUE, echo=FALSE}
smpl_point_rast <- here("data-raw", "imgs", "sample-point-rasters.png")

if (!file.exists(smpl_point_rast)) {
  "http://www.aisgzk.kz/aisgzk/Proxy/aisgzkZemPnt2/MapServer/export?dpi=96&transparent=true&format=png8&layers=show%3A147&bbox=7942828.607889999%2C6208598.464414631%2C8243720.20321%2C6512657.339685368&bboxSR=3857&imageSR=3857&size=760%2C768&_ts=1616703947841&layerDefs=147%3AShapeArea%20%3C%20748174&f=image" %>% 
  download.file(smpl_point_rast)
}

knitr::include_graphics(smpl_point_rast)
```

One ways exist, however! And we will explore it below. Here, we have an overview of what are the overall possibilities of the server and what is present there:

-  `http://www.aisgzk.kz/aisgzk/Proxy/{......}/MapServer` the root link

Available servers for `{......}` are: 

- `aisgzkZem2` - Polygons with the plots. Contains many layers here: [http://www.aisgzk.kz/aisgzk/Proxy/aisgzkZem2/MapServer?f=json](http://www.aisgzk.kz/aisgzk/Proxy/aisgzkZem2/MapServer?f=json);

    -  [Call example](http://www.aisgzk.kz/aisgzk/Proxy/aisgzkZem2/MapServer/export?dpi=96&transparent=true&format=png8&layers=show%3A147&bbox=7942828.607889999%2C6157485.119002739%2C8243720.20321%2C6563770.68509726&bboxSR=3857&imageSR=3857&size=551%2C744&_ts=1616779541796&layerDefs=147%3AShape_Area%20%3E%3D%201031965&f=image) where rasters of polygons is returned.

-  `aisgzkZemPnt2` - points for plots with the same layers as polygons. Info is here:  [http://www.aisgzk.kz/aisgzk/Proxy/aisgzkZemPnt2/MapServer?f=json](http://www.aisgzk.kz/aisgzk/Proxy/aisgzkZemPnt2/MapServer?f=json);

    -  [Call example](http://www.aisgzk.kz/aisgzk/Proxy/aisgzkZemPnt2/MapServer/export?dpi=96&transparent=true&format=png8&layers=show%3A147&bbox=7942828.607889999%2C6157485.119002739%2C8243720.20321%2C6563770.68509726&bboxSR=3857&imageSR=3857&size=551%2C744&_ts=1616779541784&layerDefs=147%3AShapeArea%20%3C%201031965&f=image) where raster image of points is returned.
    
-  `aisgzkEZKK` - Seems to be a map server for the regional boundaries, roads and extents.

-  `aisgzkTopo2` - map server that in fact returns rayon boundaries. [Call example](http://www.aisgzk.kz/aisgzk/Proxy/aisgzkTopo2/MapServer/34/query?f=json&where=RAION%20%3D%20%2709134%27&returnGeometry=true&spatialRel=esriSpatialRelIntersects&outSR=3857)


Extra info could be received from javastript code underlining the webpage [http://www.aisgzk.kz/aisgzk/js/maps.js](http://www.aisgzk.kz/aisgzk/js/maps.js)


# Logic of data extraction:


To extract data from the web, we need to follow the cadastre number and break the hierarchy of the cadastre number into the pieces. 

Generic example of a cadastre number is: `09-134-011-132`, where 

-  `09` is the `oblast`;

-  `134` is the `rayon` code;

-  `011` is the `kvartal`;

-  `132` plot id in kvartal` (Could also have 4-5 digits);

__PROBLEM!!!__ There are some rayons, where rayon code of the plots does not match the rayon code of the polygon... So in he rayon `323` leads to recieving plots with cadastre code 03-050-XXXXXX. THis requires careful re-harvesting of plots in each rayon. 


The problem is that we can only extract all ids of oblast and rayon, but not kvartals and definitely not plots. However, there is a [search form](http://www.aisgzk.kz/aisgzk/Proxy/aisgzkZem2/MapServer/find), where we can search by incomplete cadastre number and get only `10` polygons with the data. Using this search form, we can `fish` all available kvartals in each rayon and then all plots in each kvartal.

Finally, as for the metadata for each plot, it could be extracted through an ajax call `/aisgzk/Index/FindObjInfoForMap` in the script [http://www.aisgzk.kz/aisgzk/js/maps.js](http://www.aisgzk.kz/aisgzk/js/maps.js).


__Overall, fishing algorithm is the following__: 


1.  Extract all oblast/rayons from [http://www.aisgzk.kz/aisgzk/Proxy/aisgzkZem2/MapServer?f=json](http://www.aisgzk.kz/aisgzk/Proxy/aisgzkZem2/MapServer?f=json);

2.  Write an algorithm, which will learn the structure of `kvartals` and `plots` in each rayon/oblast using query form [http://www.aisgzk.kz/aisgzk/Proxy/aisgzkZem2/MapServer/find](http://www.aisgzk.kz/aisgzk/Proxy/aisgzkZem2/MapServer/find). Download all corresponding polygons in `json` at least. This is only possible up to 10 polygons at the same time, so it may take some time. 

3.  Extract metadata for each polygon by its cadastre ID number from the website's using `FindObjInfoForMap` ajax request in the underlining JS of the website here: [http://www.aisgzk.kz/aisgzk/js/maps.js](http://www.aisgzk.kz/aisgzk/js/maps.js).  


## Step 1. Gather oblast and rayon indexes

Below, we generate oblast/rayon index. 

```{r step-1, eval=TRUE, echo=FALSE}
index_clean_csv_file <-
  here("data-clean", "01-layers-index-clean.csv")

if (!file.exists(index_clean_csv_file)) {
  index_clean_json_file <-
    here("data-clean", "01-layers-index-raw.json")
  
  try({
    "http://www.aisgzk.kz/aisgzk/Proxy/aisgzkZem2/MapServer?f=json" %>%
      download.file(index_clean_json_file)
  })
  
  index_raw <- jsonlite::fromJSON(index_clean_json_file)
  
  index_kaz_adm <-
    index_raw$layers %>%
    as_tibble() %>%
    select(id, name) %>%
    separate(name, c("code", "obl_id", "rayon_id"), remove = FALSE) %>%
    select(id, name, obl_id, rayon_id) %>%
    mutate(
      date_harvested = Sys.Date(),
      spatial_refernce = as.character(index_raw$spatialReference$wkid)
    )
  
  index_kaz_adm %>%
    write_csv(index_clean_csv_file)
  
}

index_kaz_adm <- read_csv(index_clean_csv_file)
glimpse(index_kaz_adm)
```


Now, we can try harvesting actual shapes of each oblast/rayon and maybe kvartal too. [http://www.aisgzk.kz/aisgzk/Proxy/aisgzkTopo2/MapServer](http://www.aisgzk.kz/aisgzk/Proxy/aisgzkTopo2/MapServer) contains API to access actual maps/polygons and other features of the Kaz cadastre data. Table below summaries all possible mar servers with the data that could be accessed. 

We are interested in IDs 32 - Oblasts, 33 - Kvartals and 34 - Rayons. 

```{r echo = FALSE}
# raw_map_data <-
# "http://www.aisgzk.kz/aisgzk/Proxy/aisgzkTopo2/MapServer?f=json" %>% 
#   jsonlite::fromJSON()
# 
# raw_map_data$layers %>% 
#   as_tibble() 
```


There is no ways how we can load all the boundaries at once and we need to go one by one through the query form here: 

Example of a simple query for rayon info:

`http://aisgzk.kz/aisgzk/Proxy/aisgzkTopo2/MapServer/34/query?text=01001&geometryType=esriGeometryPolygon&spatialRel=esriSpatialRelIntersects&outFields=*&returnGeometry=true&returnZ=true&returnM=true&f=pjson`


After we learn what is the object id, it is also possible to access object-specif page through a direct link: `http://aisgzk.kz/aisgzk/Proxy/aisgzkTopo2/MapServer/34/276` 

Let us try harvesting rayon shapes one by one.


### 1. Call that return geometry of one rayon.

```{r geom_1_ray, eval=TRUE, echo=FALSE}
# ray_geom_call <- "http://aisgzk.kz/aisgzk/Proxy/aisgzkTopo2/MapServer/34?f=pjson"
# 
# ray_layerInfo <- jsonlite::fromJSON(ray_geom_call)
# ray_geomType <- ray_layerInfo$geometryType
# 
# ray_call_base <- 
#   str_c(
#     "http://aisgzk.kz/aisgzk/Proxy/aisgzkTopo2/MapServer/34/query?",
#     "text={obl_ray}", 
#     "&geometryType=esriGeometryPolygon", 
#     "&spatialRel=esriSpatialRelIntersects",
#     "&outFields=*&returnGeometry=true&returnZ=false&returnM=false&f=pjson"
#   )
# 
# set.seed(1221)
# random_obl_ray <- 
#   index_kaz_adm %>%
#   sample_n(1) %>%
#   mutate(obl_ray = str_c(obl_id, rayon_id)) %>%
#   pull(obl_ray)
# 
# random_obl_ray_call <- glue::glue(ray_call_base, obl_ray = random_obl_ray)
# random_obl_ray_call
```


### 2. Extracting rayon geometries

We extract here actual polygon and features of oblasts and converting it into an sf object with attributes. We also transliterate all the cyclic letters to simplify working in R.

#### 2.1 Extracting poly info


```{r 2.1-poly-info, eval=TRUE, echo=FALSE}
# raw_responce <- 
#   random_obl_ray_call %>%  
#   RCurl::getURL() 
# 
# raw_json <- 
#   raw_responce %>% 
#   jsonlite::fromJSON(simplifyVector = FALSE, simplifyDataFrame = FALSE)
# 
# str(raw_json, max.level = 1)
```


#### 2.2 Converting features and attributes to an `sf`


```{r 2.2-raw-jsons, eval=TRUE, echo=FALSE}
# Feature to polygon converter function
#
features_to_sf_polygon <- function(features) {
  ring2matrix <- function(ring) do.call(rbind, lapply(ring, unlist))
  rings2multipoly <- function(rings)
    sf::st_multipolygon(list(lapply(rings, ring2matrix)))

  getGeometry <- function(feature) {
    if (is.null(unlist(feature$geometry$rings))) {
      sf::st_multipolygon()
    } else
      rings2multipoly(feature$geometry$rings)
  }

  sf::st_sfc(lapply(features, getGeometry))
}


# raw_json_features <- 
#   raw_json$features %>% 
#   features_to_sf_polygon()
# 
# raw_json_features %>% 
#   ggplot() + 
#   geom_sf(alpha = 0.4)

```


#### 2.3 Gathering and transliterating attributes


```{r 2.3-index-conversion, eval=TRUE, echo=FALSE}
drop_nonalphanum <- 
  . %>%
  stringi::stri_trans_general("Latin-ASCII" ) %>% 
  stringi::stri_trans_general("Any-Hex/Unicode" )  %>%
  str_replace_all("U\\+02B9", "") %>%
  stringi::stri_trans_general("Hex-Any/Unicode") %>%
  str_replace_all("[^[:alnum:]_]", "")
  
get_single_attribute <- function(feat_attrs) {
  names(feat_attrs) <-
    stringi::stri_trans_general(names(feat_attrs), "Cyrl-Latn") %>%
    drop_nonalphanum
  
  feat_attrs %>%
    imap(~ {
      if (!is.numeric(.x) && !is.null(.x)) {
        if (str_detect(.y, "kaz"))
        {
          .x <- stringi::stri_trans_general(.x, "Kazakh-Latin/BGN") %>%
            stringi::stri_trans_general("Latin-ASCII") %>%
            drop_nonalphanum
        } else if (str_detect(.x, "[\u0400-\U052F]"))
        {
          .x <-
            stringi::stri_trans_general(.x, "Cyrillic-Latin/BGN") %>%
            stringi::stri_trans_general("Latin-ASCII") %>%
            drop_nonalphanum
        }
      } else if (is.null(.x)) {
        .x <- NA_character_
      }
      return(.x)
    }) %>%
    as_tibble() %>%
    janitor::clean_names()
}

get_all_attributes <- function(features, place = "attributes") {
  features %>% 
    map(place) %>% 
    map_dfr(get_single_attribute)
}

# raw_json$features[[1]]$attributes %>% 
#   get_single_attribute() %>% 
#   glimpse()

# raw_json$features %>% 
#   get_all_attributes %>% 
#   glimpse()

```


### 3. Getting all poly geometries in raw and saving them 


```{r 3.starts , eval=TRUE, echo=FALSE}

if (force_reharvest) {
  
rayons_raw_file <- 
  here(
    "data-raw",
    str_c(
      "02-1-ray-raw-json-responces-",
      Sys.time() %>% str_replace_all("[^[:alnum:]]", "-"),
      ".rds"
    )
  )

  handlers(list(
    handler_progress(
      format   = ":spin :current/:total [:bar] :percent in :elapsed ETA: :eta",
      width    = 60,
      complete = "+"
    )
  ))
  
  
  rayons_raw <-
    with_progress({
      p <- progressor(steps = nrow(index_kaz_adm))
      index_kaz_adm %>%
        mutate(
          obl_ray = str_c(obl_id, rayon_id),
          ray_url_call = glue::glue(ray_call_base)
        ) %>%
        # sample_n(10) %>%
        mutate(reponce_raw = map(ray_url_call, function(.x) {
          p()
          RCurl::getURL(.x)
        }))
    })
  
  write_rds(rayons_raw, rayons_raw_file, compress = "gz")
}


rayons_raw <-
  read_rds(here(
    "data-raw",
    "02-1-ray-raw-json-responces-2021-03-27-19-11-15.rds"
  ))

glimpse(rayons_raw)
```


### 4. Cleaning rayon polygons and saving them 


```{r 4starts , eval=TRUE, echo=FALSE}
if (force_reharvest) {
  rayons_clean_file <-
    here(
      "data-clean",
      "02-rayon-shapes",
      str_c(
        "kaz-rayons-shapes_",
        Sys.time() %>% str_replace_all("[^[:alnum:]]", "-")
      )
    )

  rayons_clean_file_shp <-
    here(
      "data-clean",
      "02-rayon-shapes",
      "shp",
      str_c(
        "kaz-rayons-shapes_",
        Sys.time() %>% str_replace_all("[^[:alnum:]]", "-")
      )
    )

  ray_geom_call <-
    "http://aisgzk.kz/aisgzk/Proxy/aisgzkTopo2/MapServer/34?f=pjson"

  ray_layerInfo <- jsonlite::fromJSON(ray_geom_call)
  # ray_geomType <- ray_layerInfo$geometryType

  rayons_clean <-
    rayons_raw %>%
    mutate(
      reponce_json =
        map(
          reponce_raw,
          ~ jsonlite::fromJSON(.x,
                               simplifyVector = FALSE,
                               simplifyDataFrame = FALSE)
        ),
      reponce_attrs =
        map(reponce_json, ~ get_all_attributes(.x$features)),
      reponce_geometries =
        map(
          reponce_json,
          ~ features_to_sf_polygon(.x$features) %>%
            sf::st_set_crs(ray_layerInfo$extent$spatialReference$latestWkid) %>%
            st_as_sf()
        )

    ) %>% {
      dta <- .
      dta %>%
        select(-reponce_geometries) %>%
        bind_cols(dta$reponce_geometries %>%
                    bind_rows() %>%
                    select(geometries = x))
    } %>%
    select(leyer_id = id,
           obl_id,
           rayon_id,
           reponce_attrs,
           geometries) %>%
    unnest(reponce_attrs) %>%
    st_as_sf()


  write_rds(rayons_clean, str_c(rayons_clean_file, ".rds"), compress = "gz")
  st_write(rayons_clean %>% select(-srvget),
           str_c(rayons_clean_file_shp, ".shp"),
           delete_dsn = T)

}

rayons_clean <-
  read_rds(here(
    "data-clean",
    "02-rayon-shapes",
    "kaz-rayons-shapes_2021-03-31-13-24-01.rds"
  ))

rayons_clean %>% 
  glimpse()

rayons_clean %>% 
  ggplot() + 
  geom_sf()

```



## Step 2.1 Fishing acual rayon code

It appears that in some rayons/API layers there are more rayons codes than it should be. It causes one to have different rayon codes. Thus, we had to harvest all rayons codes manually in each api leyer/rayon.

It is done in a separate file

```{r, echo=FALSE}
rayons_index <-
  read_rds(here(
    "data-clean/03-rayon-indexes/obl-03.rds"
  ))

rayons_index %>% 
  glimpse()

```




## Step 2. Fishing possible codes for kvartals


Kvartal boundary has a 3 digits code in cadastre for each rayon. What we will do, is going one by one digits through all sub0codes for each rayon to identify number of records displayed for each digit.

Fishing all existing kvartals in three rayons of interest:

-  03_323 - Kegen, 
-  03_044 - Enbekshikazakh and 
-  03_050 - Raiymbek


In future, this part could be re-written to fish for all kvartals in entire oblast or country.


```{r step2.1, echo=FALSE}
kvartal_index <-
  read_rds(here(
    "data-clean/04-kvartal-indexes/obl-03.rds"
  ))

kvartal_index %>% 
  glimpse()
```




### 2.2 Collecting kvartals geometries in the rayons of interest

TBD

Key requests that can help:

-  [http://www.aisgzk.kz/aisgzk/Proxy/aisgzkTopo2/MapServer/identify?f=json&tolerance=1&returnGeometry=true&returnFieldName=false&returnUnformattedValues=false&imageDisplay=159%2C769%2C96&geometry=%7B%22x%22%3A6786326.981226891%2C%22y%22%3A6367444.79994962%7D&geometryType=esriGeometryPoint&sr=3857&mapExtent=6775864.048522184%2C5952851.09152559%2C6879839.442275213%2C6455725.794645589&layers=all%3A33%2C34%2C38](http://www.aisgzk.kz/aisgzk/Proxy/aisgzkTopo2/MapServer/identify?f=json&tolerance=1&returnGeometry=true&returnFieldName=false&returnUnformattedValues=false&imageDisplay=159%2C769%2C96&geometry=%7B%22x%22%3A6786326.981226891%2C%22y%22%3A6367444.79994962%7D&geometryType=esriGeometryPoint&sr=3857&mapExtent=6775864.048522184%2C5952851.09152559%2C6879839.442275213%2C6455725.794645589&layers=all%3A33%2C34%2C38)

-  [http://www.aisgzk.kz/aisgzk/Proxy/aisgzkTopo2/MapServer/identify?f=json&tolerance=1&returnGeometry=true&returnFieldName=false&returnUnformattedValues=false&imageDisplay=159%2C769%2C96&geometry=%7B%22x%22%3A7021089.033788763%2C%22y%22%3A6027399.48704663%7D&geometryType=esriGeometryPoint&sr=3857&mapExtent=6962888.970618828%2C5785444.168250272%2C7066864.364371858%2C6288318.871370271&layers=all%3A33%2C34%2C38](http://www.aisgzk.kz/aisgzk/Proxy/aisgzkTopo2/MapServer/identify?f=json&tolerance=1&returnGeometry=true&returnFieldName=false&returnUnformattedValues=false&imageDisplay=159%2C769%2C96&geometry=%7B%22x%22%3A7021089.033788763%2C%22y%22%3A6027399.48704663%7D&geometryType=esriGeometryPoint&sr=3857&mapExtent=6962888.970618828%2C5785444.168250272%2C7066864.364371858%2C6288318.871370271&layers=all%3A33%2C34%2C38)




## Step 3. Getting shapes of the plots


We develop a set of functions in a separate files, as well as scripts which are used to extract plots data. Here we only summarize the results.

Overall, process of extracting plots geometries looks in a following way:

Each geometry has a unique cadastre code XX-YYY-ZZZ-SSSF

Where:

XX - Oblast
YYY - Rayon 
ZZZ - Kvartal
SSS - plot
F   - Extra digit for a plot ID present in some SSS plots. 


We can make a search request for XX-YYY and the server will only return 10 first records for us. Therefore, in order to get all plots in each kvartal, we need to:

1. Identify YYY of a rayon and ZZZ for this kvartal.
2. Identify first valid "S" digits, which have plots in each kvartal
3. Search each second S digit and harvest all 10 third "S" digits.
4. Harvest extra F digits if at step 3 we got a message that limit for request transfer was reached.  


This is all done in a separate file. 


### 3.5 Checking the plots

```{r, echo=FALSE}
poly_dta <- 
  read_rds("data-clean/10-sarah-request/kaz-akmol-3-ray-1ha-clean.rds") %>% 
  st_as_sf() %>% 
  left_join(rayons_index %>% select(leyer_id, obl_rus, raj_rus ), 
            c("layerId"  = "leyer_id"))

glimpse(poly_dta)

```

Plotting all 1 ha and larger plots:

```{r, echo=FALSE}
out_gg <- 
  poly_dta %>% 
  filter(Shape_Area > 10000) %>% 
  ggplot() + 
  aes(fill = raj_rus) + 
  geom_sf() + 
  theme_minimal()


ggsave("data-clean/10-sarah-request/3-ray-example.pdf", width = 30, height = 15)

out_gg

```


# Interactive map


```{r out.width="100%", echo=FALSE}
library(mapview)

# rayon_bounds <- 
#   rayons_clean %>%
#     filter(leyer_id %in% unique(poly_dta$layerId))

poly_dta %>% 
  filter(Shape_Area > 10000) %>% 
  select(cadastre_id, obl_rus, raj_rus, CATEGORY_RUS:NAZV) %>% 
mapview(
    map.type =
      c(
        "CartoDB.Positron",
        "Esri.WorldImagery",
        "CartoDB.DarkMatter",
        "OpenStreetMap",
        "OpenTopoMap"
      ),
    zcol = "raj_rus",
    burst = TRUE
  )

poly_dta %>%
  filter(actual_rayon_id != rayon_id) %>%
  st_drop_geometry() %>%
  as_tibble() %>%
  distinct(rayon_id, actual_rayon_id, CATEGORY_RUS)

poly_dta %>%
  filter(actual_rayon_id != rayon_id) %>%
  st_drop_geometry() %>%
  as_tibble() %>%
  distinct(rayon_id, actual_rayon_id, PRAVO_RUS)
```   


## Step 4. Getting metadata for existing cadatre codes

Harvested in a separate document. Below is the summary of the main features of the dataset.

```{r, echo=FALSE}
parced_metadata_clean <- 
  read_rds(here("data-clean","06-plots-metadata-clean.rds"))

parced_metadata_clean %>% glimpse()

parced_metadata_clean %>% count(rus_predostavlennoe_pravo, sort = T)

parced_metadata_clean %>% count(rus_kategoria_zemel, sort = T)

parced_metadata_clean %>% count(rus_celevoe_naznacenie, sort = T)

parced_metadata_clean %>% count(rus_zemlepolzovateli, sort = T)

parced_metadata_clean %>% count(rus_ogranicenia, sort = T)

```
